# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smartbiz Payroll API",
        "description": "API documentation for Smartbiz Payroll",
        "contact": {
            "email": "adisayoga@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://smartbiz-payroll.test/api/v1",
            "description": "Smartbiz Payroll"
        }
    ],
    "paths": {
...
        "/payroll-groups": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get all payroll group per page",
                "operationId": "ac774a2bcabce2389f27ae78f70f811e",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per-page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "{field: string, operator?: string, value: any}[]"
                        }
                    },
                    {
                        "name": "sorts",
                        "in": "query",
                        "schema": {
                            "type": "string[]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll group",
                "operationId": "521ce86b901d40072a9584dca1d91af8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-groups/list-cbx": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "List all payroll groups for combob box",
                "operationId": "ff9e91b0bb396517571591e28242aa63",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-groups/{payrollGroupId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll group",
                "operationId": "20b27c9e00a22e571a5d90f3ef2b6e17",
                "parameters": [
                    {
                        "name": "payrollGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll group",
                "operationId": "5a5763d84e3abe62fd4fc2793212d8f7",
                "parameters": [
                    {
                        "name": "payrollGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll group",
                "operationId": "a8c45f42f9e964e1e2e01ddab6ac889b",
                "parameters": [
                    {
                        "name": "payrollGroupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/tracking-history/{name}/{id}": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "List date tracking history per item",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number | string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/employees/list-cbx": {
            "get": {
                "tags": [
                    "Personal"
                ],
                "summary": "Get all employees for combobox",
                "operationId": "543b06a06b7b16b71c9fba4e2610bb8e",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/departments/list-cbx": {
            "get": {
                "tags": [
                    "WorkStructure"
                ],
                "summary": "Get all departments/work units for combobox",
                "operationId": "dcf15bf77fee5d7ae452e3f856a39ef2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/grades/list-cbx": {
            "get": {
                "tags": [
                    "WorkStructure"
                ],
                "summary": "Get all employee grades for combobox",
                "operationId": "03e35a6ca0bb21c6257144499dfbf747",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer [token]",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}